requirements.txt
aniso8601==7.0.0
asgiref==3.3.0
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.1
colorama==0.4.4
Django==3.1.3
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
dlib==19.19.99
face-recognition==1.3.0
face-recognition-models==0.3.0
gql==2.0.0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.22
mysqlclient==2.0.1
numpy==1.18.2
Pillow==7.0.0
promise==2.3
PyJWT==1.7.1
pylint==2.6.0
pytz==2020.4
requests==2.25.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Unidecode==1.1.1
urllib3==1.26.2
wrapt==1.12.1
